Создана папка pages. Создан компонент страницы Home, в этот компонент вынесена основная логика компонента App. Добавлен Preloader.
