--- src/components/app/app.module.css
@@ -1,13 +1,0 @@
-.content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.main {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    padding: 16px 20px;
-    gap: 40px;
-}
